% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempoR4SPISE2024.R
\docType{package}
\name{tempoR4SPISE2024-package}
\alias{tempoR4SPISE2024-package}
\alias{tempoR4SPISE2024}
\title{A set of Programs, Vignettes,
and Data for the SPISE2024 Advanced Workshop}
\description{
A set of Programs, Vignettes,
and Data for the SPISE2024 Advanced Workshop
(Hanoi Vietnam): 07/24/2024 to 07/25/2024.
}
\details{
\code{tempoR4SPISE2024}:
Provides a set of Programs, Vignettes,
and Data for the SPISE2024 Advanced Workshop
(Hanoi Vietnam).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/HerveAbdi/tempoR4SPISE2024}
  \item \url{https://herveabdi.github.io/tempoR4SPISE2024/}
  \item Report bugs at \url{http://github.com/HerveAbdi/tempoR4SPISE2024/issues}
}

}
\author{
\strong{Maintainer}: Hervé Abdi \email{herve@utdallas.edu} (\href{https://orcid.org/0000-0002-9522-1978}{ORCID})

Other contributors:
\itemize{
  \item Vincent Guillemot \email{vincent.guillemot@pasteur.fr} (\href{https://orcid.org/0000-0002-7421-0655}{ORCID}) [contributor]
  \item Ju-Chi Yu \email{Ju-Chi.Yu@camh.org} (\href{https://orcid.org/0000-0002-6360-1861}{ORCID}) [contributor]
  \item Soudeh A. Khoubrouy \email{SA.Khoubrouy@utdallas.edu} (\href{https://orcid.org/0000-0002-2656-7756}{ORCID}) [contributor]
}

}
